+++ src/main/java/com/hyperprotect/mixin/intercept/interaction/SimpleBlockInteractionGate.java
@@ -292,7 +292,10 @@ public abstract class SimpleBlockInteractionGate {
                                 }
                                 var state = blockType.getState();
                                 if (state != null) {
-                                    System.getProperties().put("hyperprotect.context.block_state", state.getId());
+                                    String stateStr = state.getStateForBlock(blockType.getId());
+                                    if (stateStr != null) {
+                                        System.getProperties().put("hyperprotect.context.block_state", stateStr);
+                                    }
                                 }
                             }
                         } catch (Exception ignored) {}